feat: add QR code PDF generator for reservation details in a separated component
