+++ src/app/[locale]/reservations/confirmed/ClientConfirmationPage.tsx
@@ -8,9 +8,8 @@ import { SerializedSession } from '@/app/[locale]/lib/stripe/types'
 import { Calendar, Ticket, Type, DollarSign } from 'lucide-react'
 import { useTranslations } from 'next-intl'
 
-import { jsPDF } from 'jspdf'
-import QRCode from 'qrcode'
 import { ReservationDetails } from '../../interfaces/interfaces'
+import { QRCodePDFGenerator } from '../../components/QRCodePDFGenerator'
 
 interface ClientConfirmationPageProps {
   reservationId?: string | null
@@ -101,53 +100,6 @@ export function ClientConfirmationPage({
     fetchReservationDetails()
   }, [reservationId])
 
-  //PDF GENERATOR
-  const generatePDF = async (reservationDetails: ReservationDetails | null) => {
-    if (!reservationDetails) return
-
-    // Crear QR con la información
-    const qrData = `
-    ${t('ReservationNumber')} ${reservationId}\n
-    `
-    const qrImage = await QRCode.toDataURL(qrData)
-
-    // Crear documento PDF
-    const doc = new jsPDF()
-    doc.setFontSize(16)
-    doc.setFont('helvetica', 'bold')
-    doc.text(`${t('ReservationDetails')}`, 20, 10)
-    // Obtener el ancho del texto para calcular el subrayado
-    const textWidth = doc.getTextWidth(`${t('ReservationDetails')}`)
-    doc.line(20, 12, 20 + textWidth, 12) // Dibujar la línea de subrayado
-    // Restablecer la fuente a normal si es necesario
-    doc.setFont('helvetica', 'normal')
-    doc.text(`${t('ReservationType')} ${reservationDetails?.visitTypeId || 'N/A'}`, 20, 20)
-    doc.text(`${t('TicketNumber')} ${reservationDetails?.numPeople || 0}`, 20, 30)
-    doc.text(
-      `${t('Date')} ${reservationDetails?.createdAt?.toDate().toLocaleDateString() || 'N/A'}`,
-      20,
-      40,
-    )
-    doc.text(
-      `${t('TotalPrice')} ${
-        reservationDetails?.totalAmount ? reservationDetails.totalAmount / 100 : '0.00'
-      } ${reservationDetails?.currency?.toUpperCase()}`,
-      20,
-      50,
-    )
-
-    doc.setTextColor(255, 0, 0) // Establecer color rojo
-    doc.text(`${t('TextScan')}`, 20, 70)
-
-    // Restablecer el color si es necesario para los siguientes textos
-    doc.setTextColor(0, 0, 0)
-
-    doc.addImage(qrImage, 'PNG', 20, 80, 100, 100) // Agregar el QR
-
-    // Descargar el PDF
-    doc.save('reservation.pdf')
-  }
-
   return (
     <div className="container mx-auto py-10 px-4">
       <div className="max-w-2xl mx-auto  ">
@@ -211,15 +163,10 @@ export function ClientConfirmationPage({
           )} */}
 
           <div className="flex flex-col md:flex-row justify-center items-center md:gap-2 gap-4">
-            <div className="space-y-4 text-center inline-flex w-full">
-              <button
-                onClick={() => generatePDF(reservationDetails)}
-                className="px-6 py-3 bg-blue-600 text-white rounded-md hover:bg-blue-700 transition-colors w-full"
-              >
-                {t('Download')}
-              </button>
-            </div>
-
+            <QRCodePDFGenerator
+              reservationId={reservationId}
+              reservationDetails={reservationDetails}
+            />
             <div className="space-y-4 text-center inline-flex w-full">
               <Link
                 href="/"
